# Desktop/bankingApp/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Banking Application","description":"API для управления банковских операций клиентов"}],"paths":{"/transfer/{userId}":{"post":{"tags":["Banking Application"],"summary":"Перевод денежных средств","description":"Выполнение перевода со счета аутентифицированного пользователя, на счёт другого пользователя.","operationId":"transferMoney","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferRequestDto"}}},"required":true},"responses":{"404":{"description":"Ошибка при поиске данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Ошибка с данными при создании пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Ошибка валидации данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Операция выполнена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}},"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}}}}},"/registration":{"post":{"tags":["Banking Application"],"summary":"Создание нового пользователя","description":"У пользователя также есть телефон и email. Телефон и или email должен быть минимум один. На “банковском счету” должна быть какая-то изначальная сумма. Также у пользователя должна быть указана дата рождения и ФИО.","operationId":"addNewClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientRequestDto"}}},"required":true},"responses":{"404":{"description":"Ошибка при поиске данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Ошибка с данными при создании пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Ошибка валидации данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Операция выполнена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}},"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}}}}},"/auth":{"post":{"tags":["Banking Application"],"summary":"Создание токена пользователя","description":"Процесс авторизации пользователя","operationId":"createAuthenticatedToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtRequest"}}},"required":true},"responses":{"404":{"description":"Ошибка при поиске данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Ошибка с данными при создании пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Ошибка валидации данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Операция выполнена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}},"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/add/{userId}":{"post":{"tags":["Banking Application"],"summary":"Добавление контактной информации","description":"Пользователь может добавить свои номер телефона и/или email, если они еще не заняты другими пользователями.","operationId":"addClientContactInformation","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactInfoRequest"}}},"required":true},"responses":{"404":{"description":"Ошибка при поиске данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Ошибка с данными при создании пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Ошибка валидации данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Операция выполнена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}},"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}}}}},"/update/{userId}":{"patch":{"tags":["Banking Application"],"summary":"Изменение контактной информации","description":"Пользователь может сменить свои номер телефона и/или email, если они еще не заняты другими пользователями.","operationId":"updateClientContactInformation","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactInfoRequest"}}},"required":true},"responses":{"404":{"description":"Ошибка при поиске данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Ошибка с данными при создании пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Ошибка валидации данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Операция выполнена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}},"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponseDto"}}}}}}},"/search":{"get":{"tags":["Banking Application"],"summary":"Поиск пользователей ","description":"Поиск выполняется по параметрам или без них, с применением сортировки","operationId":"getClients","parameters":[{"name":"birthDay","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"phoneNumber","in":"query","required":false,"schema":{"type":"string"}},{"name":"fullName","in":"query","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"from","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"404":{"description":"Ошибка при поиске данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Ошибка с данными при создании пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Ошибка валидации данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Операция выполнена","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientResponseDto"}}}}},"201":{"description":"Пользователь создан","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientResponseDto"}}}}}}}},"/delete/{userId}":{"delete":{"tags":["Banking Application"],"summary":"Удаление контактной информации","description":"Удаление телефона/эл.почты, если они у него не единственные","operationId":"removeClientContactInformation","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactInfoRequest"}}},"required":true},"responses":{"404":{"description":"Ошибка при поиске данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Ошибка с данными при создании пользователя","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"Unauthorized","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Ошибка валидации данных","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Операция выполнена"},"201":{"description":"Пользователь создан"},"204":{"description":"No Content"}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"errors":{"type":"array","description":"Список стектрейсов или описания ошибок","items":{"type":"string","description":"Список стектрейсов или описания ошибок"}},"message":{"type":"string","description":"Сообщение об ошибке"},"reason":{"type":"string","description":"Общее описание причины ошибки"},"status":{"type":"string","description":"Код статуса HTTP-ответа","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 EARLY_HINTS","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"timestamp":{"type":"string","description":"Дата и время когда произошла ошибка (в формате \"yyyy-MM-dd HH:mm:ss\")","format":"date-time"}},"description":"Сведения об ошибке"},"TransferRequestDto":{"required":["receiverId","sum"],"type":"object","properties":{"receiverId":{"type":"integer","description":"Идентификатор получателя","format":"int64"},"sum":{"type":"number","description":"Сумма перевода","format":"double"}},"description":"DTO для перевода денежных средств"},"ClientResponseDto":{"type":"object","properties":{"id":{"type":"integer","description":"Уникальный идентификатор клиента","format":"int64"},"surname":{"type":"string","description":"Фамилия клиента"},"name":{"type":"string","description":"Имя клиента"},"patronymic":{"type":"string","description":"Отчество клиента"},"birthDay":{"type":"string","description":"Дата рождения клиента","format":"date"},"phoneNumbers":{"type":"array","description":"Список телефонных номеров клиента","items":{"type":"string","description":"Список телефонных номеров клиента"}},"emails":{"type":"array","description":"Список электронных адресов клиента","items":{"type":"string","description":"Список электронных адресов клиента"}},"balance":{"type":"number","description":"Баланс клиента","format":"double"},"login":{"type":"string","description":"Логин клиента"}},"description":"DTO ответа клиента"},"ClientRequestDto":{"required":["birthDay","confirmPassword","email","login","name","password","patronymic","phoneNumber","startDeposit","surname"],"type":"object","properties":{"surname":{"type":"string","description":"Фамилия клиента"},"name":{"type":"string","description":"Имя клиента"},"patronymic":{"type":"string","description":"Отчество клиента"},"birthDay":{"type":"string","description":"Дата рождения клиента","format":"date"},"phoneNumber":{"maxLength":10,"minLength":10,"type":"string","description":"Номер телефона клиента"},"email":{"type":"string","description":"Email клиента"},"startDeposit":{"minimum":0.0,"type":"number","description":"Начальный депозит клиента","format":"double"},"login":{"type":"string","description":"Логин клиента"},"password":{"type":"string","description":"Пароль клиента"},"confirmPassword":{"type":"string","description":"Проверка пароля клиента"}},"description":"DTO для создания нового клиента"},"JwtRequest":{"required":["login","password"],"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}}},"JwtResponse":{"type":"object","properties":{"token":{"type":"string"}}},"ContactInfoRequest":{"required":["status"],"type":"object","properties":{"phone":{"maxLength":10,"minLength":10,"type":"string","description":"Телефонный номер для добавления или удаления","nullable":true},"changePhone":{"maxLength":10,"minLength":10,"type":"string","description":"Изменяемый телефонный номер","nullable":true},"email":{"type":"string","description":"Адрес электронной почты для добавления или удаления","nullable":true},"changeEmail":{"type":"string","description":"Изменяемый адрес электронной почты","nullable":true},"status":{"type":"string","description":"Статус операции: ADD (добавление), REPLACE (замена), DELETE (удаление)","enum":["DELETE","ADD","REPLACE"]}},"description":"Контактные данные для удаления/добавления/изменения"}}}}